+++ devicelayer/include/deviceLayer/IDeviceLayer.h
@@ -379,14 +379,14 @@ public:
   /// and service processor will be reinitialized.
   /// @param[in] timeout the operation will be aborted if time exceeds given timeout
   ///
-  virtual void reinitDeviceInstance(int device, bool masterMinionOnly, std::chrono::milliseconds timeout);
+  virtual void reinitDeviceInstance(int device, bool masterMinionOnly, std::chrono::milliseconds timeout) = 0;
 
   /// \brief Hints the underlying device that we don't have anything to process, so it could potentially go idle / low
   /// power consumption.
   ///
   /// @param[in] device the device to be hinted
   ///
-  virtual void hintInactivity(int device);
+  virtual void hintInactivity(int device) = 0;
 
   /// \brief Evaluates the Peer-to-Peer DMA compatibility of given devices.
   ///
@@ -395,7 +395,7 @@ public:
   ///
   /// @returns true if deviceA and deviceB are compatible for P2P DMA, false otherwise
   ///
-  virtual bool checkP2pDmaCompatibility(int deviceA, int deviceB) const;
+  virtual bool checkP2pDmaCompatibility(int deviceA, int deviceB) const = 0;
 };
 
 class IDeviceLayer : public IDeviceAsync, public IDeviceSync {