+++ devicelayer/src/DevicePcie.h
@@ -57,6 +57,7 @@ public:
   std::string getDeviceAttribute(int device, std::string relAttrPath) const override;
   void clearDeviceAttributes(int device, std::string relGroupPath) const override;
   void reinitDeviceInstance(int device, bool masterMinionOnly, std::chrono::milliseconds timeout) override;
+  void hintInactivity(int) override;
   bool checkP2pDmaCompatibility(int deviceA, int deviceB) const override;
 
 private: