+++ devicelayer/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixed
 - [SW-18867] Fix compilation with GCC 11
+- [SW-19102] Make IDeviceSYnc a pure interface (breaking change)
 ### Security
 
 ## [2.2.0] - 2023-08-03